bloodwork-dashboard/requirements.txt
fastapi==0.108.0
uvicorn==0.27.0
supabase==1.0.3
python-dotenv==1.0.0
openai==1.55.3
pydantic==2.8.0
python-multipart==0.0.6
requests==2.32.3